Build - Added basic build system and asy way to include
