+++ examples/pose/pose_local_pose_anim_blend.c
@@ -17,7 +17,7 @@
 
 #include "../common/boilerplate_main.h"
 
-#include "../../src/skeleton.c"
+#include "skeleton.h"
 
 enum Anim {
   IDLE,
