+++ src/bone_mask.h
